+++ engine/server/engine/centralized_logs/loki_log_pipeline.go
@@ -2,29 +2,14 @@ package centralized_logs
 
 import (
 	"fmt"
-	"github.com/kurtosis-tech/stacktrace"
 )
 
 type lokiLogPipeline struct {
 	lineFilters []*LokiLineFilter
 }
 
-func NewLokiLogPipeline(lineFilters []*LokiLineFilter) (*lokiLogPipeline, error) {
-	for lineFilterIndex, lineFilter := range lineFilters{
-		lineFilterOperator := lineFilter.GetOperator()
-		if !lineFilterOperator.IsDefined(){
-			lineFilterPosition := lineFilterIndex + 1
-			return nil, stacktrace.NewError(
-				"New Loki log line with line filters '%+v' can't be created because the operator '%v' is not defined for the line filter in position '%v' and with text '%v'",
-				lineFilters,
-				lineFilterOperator,
-				lineFilterPosition,
-				lineFilter.GetText(),
-			)
-		}
-	}
-
-	return &lokiLogPipeline{lineFilters: lineFilters}, nil
+func NewLokiLogPipeline(lineFilters []*LokiLineFilter) *lokiLogPipeline {
+	return &lokiLogPipeline{lineFilters: lineFilters}
 }
 
 func (logPipeline *lokiLogPipeline) PipeLineStringify() string{